refactor(test): move tests to existing package
